Lighting: Split into files & refactor
Each of the core parts of the lighting arduino
- LED control
- Main modes (red alert, warp etc)
- Special effects (damage flicker)
are now in seperate files.
This is rather like how the upper engineer works, but with fewer files
(technically I could have gone with a seperate file per mode/effect but nah)

Hopefully this makes it easier to read/understand.

(Also C++11 because it turns out arduino uses that!)